pyproject.toml
[project]
name = "doggy-neural-glitch"
version = "0.1.0"
description = "Demoscene-inspired audio-to-visual glitch art generator"
requires-python = ">=3.10"
dependencies = [
    "numpy>=1.24.0",
    "Pillow>=10.0.0",
    "librosa>=0.10.0",
    "soundfile>=0.12.0",
    "torch>=2.0.0",
    "tqdm>=4.65.0",
    "scipy>=1.10.0",
]
